fixed issue #26: compile with a 64bit environment

+++ TinyJS_Threading.cpp
@@ -30,6 +30,7 @@
 #include "TinyJS_Threading.h"
 #include <exception>
 #include <cstdio>
+#include <cstddef>
 
 #undef HAVE_THREADING
 #if !defined(NO_THREADING) && !defined(HAVE_CUSTOM_THREADING_IMPL)
@@ -187,15 +188,15 @@ public:
 		if(Wait) {
 			pthread_join(thread, &retvar);
 		}
-		return (int)retvar;
+		return (int32_t)((ptrdiff_t)retvar);
 	}
-	int retValue() { return (int)retvar; }
+	int retValue() { return (int32_t)((ptrdiff_t)retvar); }
 	bool isActiv() { return activ; }
 	bool isRunning() { return running; }
 	bool isStarted() { return started; }
 	static void *ThreadFnc(CScriptThread_impl *This) {
 		This->running = This->started = true;
-		This->retvar = (void*)This->This->ThreadFnc();
+		This->retvar = (void*)((ptrdiff_t)This->This->ThreadFnc());
 		This->running = false;
 		This->This->ThreadFncFinished();
 		return This->retvar;
